contrib/modelhub/requirements.txt
keras-applications
huggingface-hub
tensorflow-hub
onnx
requests
pytest
torch-model-archiver